+++ repository/src/main/java/com/bhaktaprogram/repository/di/DatabaseComponent.kt
@@ -1,12 +1,8 @@
 package com.bhaktaprogram.repository.di
 
 import android.content.Context
-import androidx.room.Room
-import com.bhaktaprogram.repository.database.AppDatabase
 import dagger.BindsInstance
 import dagger.Component
-import dagger.Module
-import dagger.Provides
 import javax.inject.Singleton
 
 @Singleton
@@ -20,22 +16,5 @@ interface DatabaseComponent {
         fun setContext(context: Context): Builder
 
         fun build(): DatabaseComponent
-    }
-}
-
-@Module
-class DatabaseModule {
-
-    companion object {
-
-        private const val DATABASE_NAME = "notes.db"
-
-        @Provides
-        @Singleton
-        fun provideDatabase(context: Context) = Room.databaseBuilder(
-            context,
-            AppDatabase::class.java,
-            DATABASE_NAME
-        ).build()
     }
 }
